Log és email küldés

+++ app/Http/Controllers/api/CategoryController.php
@@ -8,6 +8,8 @@ use App\Models\Category;
 use App\Http\Requests\CategoryRequest;
 use App\Traits\ResponseTrait;
 use App\Services\CategoryService;
+use Illuminate\Support\Facades\Gate;
+use App\Models\User;
 
 class CategoryController extends Controller {
     
@@ -21,16 +23,20 @@ class CategoryController extends Controller {
 
     public function getCategories() {
 
+        Gate::authorize( "viewAny", Category::class );
+
         $categories = Category::all();
 
         return $this->sendResponse( $categories );
     }
 
-    public function createCategory( CategoryRequest $request ) {
+    public function createCategory( CategoryRequest $request, User $user ) {
+
+        Gate::authorize( "create", Category::class );
 
         $validated = $request->validated();
         
-        return $this->categoryService->create( $validated );
+        return $this->categoryService->create( $validated, auth()->user() );
     }
 
     public function updateCategory( CategoryRequest $request, Category $category ) {
@@ -42,6 +48,8 @@ class CategoryController extends Controller {
 
     public function destroyCategory( Category $category ) {
 
+        Gate::authorize( "delete", $category );
+        
         return $this->categoryService->delete( $category );
     }
 }
